refactor: rename worker request handler files
