+++ jina/serve/runtimes/worker/__init__.py
@@ -13,7 +13,7 @@ from jina.proto import jina_pb2, jina_pb2_grpc
 from jina.serve.instrumentation import MetricsTimer
 from jina.serve.runtimes.asyncio import AsyncNewLoopRuntime
 from jina.serve.runtimes.helper import _get_grpc_server_options
-from jina.serve.runtimes.request_handlers.data_request_handler import WorkerRequestHandler
+from jina.serve.runtimes.request_handlers.worker_request_handler import WorkerRequestHandler
 from jina.types.request.data import DataRequest
 
 if TYPE_CHECKING: # pragma: no cover
@@ -98,7 +98,7 @@ class WorkerRuntime(AsyncNewLoopRuntime, ABC):
         # Keep this initialization order
         # otherwise readiness check is not valid
         # The WorkerRequestHandler needs to be started BEFORE the grpc server
-        self._data_request_handler = WorkerRequestHandler(
+        self._worker_request_handler = WorkerRequestHandler(
             self.args,
             self.logger,
             self.metrics_registry,
@@ -163,7 +163,7 @@ class WorkerRuntime(AsyncNewLoopRuntime, ABC):
         """Close the data request handler"""
         self._health_servicer.enter_graceful_shutdown()
         await self.async_cancel()
-        self._data_request_handler.close()
+        self._worker_request_handler.close()
 
     async def process_single_data(self, request: DataRequest, context) -> DataRequest:
         """
@@ -186,7 +186,7 @@ class WorkerRuntime(AsyncNewLoopRuntime, ABC):
         self.logger.debug('got an endpoint discovery request')
         endpointsProto = jina_pb2.EndpointsProto()
         endpointsProto.endpoints.extend(
-            list(self._data_request_handler._executor.requests.keys())
+            list(self._worker_request_handler._executor.requests.keys())
         )
         return endpointsProto
 
@@ -216,7 +216,7 @@ class WorkerRuntime(AsyncNewLoopRuntime, ABC):
                 tracing_context = WorkerRuntime._extract_tracing_context(
                     context.invocation_metadata()
                 )
-                result = await self._data_request_handler.handle(
+                result = await self._worker_request_handler.handle(
                     requests=requests, tracing_context=tracing_context
                 )
                 if self._successful_requests_metrics:
@@ -235,7 +235,7 @@ class WorkerRuntime(AsyncNewLoopRuntime, ABC):
                     exc_info=not self.args.quiet_error,
                 )
 
-                requests[0].add_exception(ex, self._data_request_handler._executor)
+                requests[0].add_exception(ex, self._worker_request_handler._executor)
                 context.set_trailing_metadata((('is-error', 'true'),))
                 if self._failed_requests_metrics:
                     self._failed_requests_metrics.inc()
